Fix browser export
