+++ packages/utils/src/npm.ts
@@ -4,14 +4,12 @@ import {
   SnapFiles,
   SnapValidationFailureReason,
   UnvalidatedSnapFiles,
-} from './types';
-import { ProgrammaticallyFixableSnapError, validateSnapShasum } from './snaps';
-import {
   assertIsNpmSnapPackageJson,
   assertIsSnapManifest,
   NpmSnapPackageJson,
   SnapManifest,
-} from './manifest';
+} from './types';
+import { ProgrammaticallyFixableSnapError, validateSnapShasum } from './snaps';
 
 export const SVG_MAX_BYTE_SIZE = 100_000;
 export const SVG_MAX_BYTE_SIZE_TEXT = `${Math.floor(
