--- packages/utils/src/test-utils/types.ts
@@ -1,1 +1,0 @@
-export type PartialOrNull<T> = { [P in keyof T]?: T[P] | undefined | null };
